// dashboard/ui/src/api/generated/models/mDPositPublication.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * MDDash Dashboard API
 * API for experiment, simulation, compute, analysis, notebook, file, and publication workflows.
 * OpenAPI spec version: 1.0.0
 */
import type { NullableString } from './nullableString.ts';
import type { PublicationFile } from './publicationFile.ts';

export interface MDPositPublication {
  metadata_file: PublicationFile;
  files: PublicationFile[];
  vre_lite_url: NullableString | null;
}