// dashboard/ui/src/api/generated/models/mDRepoPublication.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * MDDash Dashboard API
 * API for experiment, simulation, compute, analysis, notebook, file, and publication workflows.
 * OpenAPI spec version: 1.0.0
 */
import type { MDRepoPublicationLinks } from './mDRepoPublicationLinks.ts';
import type { NullableString } from './nullableString.ts';

export interface MDRepoPublication {
  id: string;
  links: MDRepoPublicationLinks;
  upload_id: string;
  upload_state: string;
  draft_url: NullableString | null;
}